Update 3.0.5
Changed the way EXP, Rewards, and other loading systems work to a per player YML system, this is in DEVELOPMENT, and isn't guarenteed to work.

Removed remains of the failed MYSQL attempts.

Sorry that I have been inactive for about a week, got tired of coding and took a break.