Update some more functions (Incomplete)
I intend to update the battles functions this weekend, but cant embark on the project until then (because of how complex battles is I need to get an understanding on it) though I will start taking a look and maybe making some changes. I do plan to work on ship types issue, and possibly super charged engines this week.